+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Upgraded to PHPStan 1.x and fixed a number of issues it detected.
 * Removed the old demo and extras files as no applicable to version 3.
 * Removed support for `Resque_Job_Creator`.
+* Removed deprecated properties and methods from `Worker` and `WorkerFactory`.
 
 ## 2.12.2 (2023-03-02)